Added custom AutoExposure algorithm
Added Mutexes on member variables Cam1_Image and Cam2_Image.
This allows the auto-exposure algorithm to run on a different thread
safely by ensuring that the latest updated frame does not change in
the process of copying it to a thread-safe variable.

Also changed default ROI to align better with the endoscope, as well
as reducing the threshold minimum difference to 100.