Fixed video recording on 3rd screen
Also fixed directories where screenshots/videos are saved. MacOS and
Apple are "special" in that the root of the program is located 3 folders
deep from where the program itself is executed. As a result, the files
were saved 3 folders in. This commit creates new folders "Screenshot"
and "Video" if they don't already exist, located where you would normally
open the program, and saves video/screenshots there.