Overhauled NIR thresholding system
NIR now uses a true color-lookup table, which automatically detects and assigns
a fixed percentage of pixel intensity values a color. This allows the image
to have dynamic contrast, making full use of the color-lookup table as opposed
to only using 1 or 2 colors. Background noise is still rendered as black, and
the value for it is hardcoded as of now. In the future, this value will
be calibrated automatically via a push of a button.