Removed binning and completely overhauled Median filter
Binning was amplifying the noise too much, and actually reduced
the signal-to-noise ratio unlike previously thought to do. Namely,
binning caused an accumulation of charge in the top right section
of the NIR camera. After discussion,  we opted to remove it.

The original median filter has been overhauled, and given its own
function as well. This new median function allows for a variable
radius, and operates in O(r) as opposed to the previous
"brute-force" algorithm, which operated in O(r^2*log(r)). The
performance boost is extremely noticeable, with radius 3 comparable
(and possibly even faster) to the original algorithm with radius 1.

The increase in speed has allowed for an increase in filtering
"power". The NIR display (with a properly-calibrated threshold)
shows almost no noise.